+++ lib/puppet/functions/extended_stdlib/decrypt_string.rb
@@ -4,7 +4,7 @@
 Puppet::Functions.create_function(:'extended_stdlib::decrypt_string') do
   # @param key The path to the file containing the cipher key.
   # @param nonce The path to the file containing the cipher nonce.
-  # @param encrypted The encrypted string to decrypt.
+  # @param encrypted The path to the file containing the encrypted string to decrypt.
   # @return String Returns the encrypted string as decrypted in plain text.
   # @example Decrypt an encrypted string and return it plain text.
   #   decrypt_string('/path/to/key.txt', '/path/to/nonce.txt', '/path/to/encrypted.txt') => foobar