+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix docstrings issues.
 - Input parameter validation for `drop`, `center`, and `log` functions.
 - Error when `intersect` and `start_with` are not supported in Ruby version.
+- Minor code optimization.
 
 # 1.3.0
 - Add functions: http_get.